refactor scale selectors to seperate files
closes #542
