+++ projects/ngx-rxtimeline/src/lib/activity/activity.selectors.ts
@@ -12,7 +12,8 @@ import { TimelineDragEvent } from '../drag/timeline-drag-event';
 import { PositionedActivity } from './positioned-activity';
 import { InverseBandScale, TimeScale } from '../scales/scale-types';
 import { Orientation } from '../core/orientation';
-import { selectBandScale, selectTimeScale } from '../scales/scale-selectors';
+import { selectBandScale } from '../scales/selectors/band-scale.selectors';
+import { selectTimeScale } from '../scales/selectors/time-scale.selectors';
 import { getInverseBandScale } from '../scales/scale-utils';
 import { MemoizedSelector } from '../store-lib/selector/memoized-selector';
 import { selectGetTypeZIndex } from '../options/selectors/type-options.selectors';
